core/detection/requirements.txt
torch==2.5.0
opencv-python==4.8.0.74
pillow==9.3.0
numpy==1.24.3
setuptools==58.1.0
wheel==0.38.4
